Updated login and register:
Login has been updated to return all user values except password upon login. Register has been updated to create a new StudentCard if an id is given that is not in the database. This does make the suto incremented keys for StudentCardId redundant though, we may need to change that.